Improve profile info drawer by showing the slug
