+++ apps/website/inertia/components/profile/profile_info_drawer.tsx
@@ -2,9 +2,7 @@ import ParticipantProfile from "#models/participant_profile";
 import { Link } from "@tuyau/inertia/react";
 import { User } from "lucide-react";
 import { Drawer, DrawerContent } from "~/components/ui/drawer";
-import { Button } from "~/components/ui/button";
-import NFC from "../credentials/nfc";
-import { useTuyau } from "~/hooks/use_tuyau";
+import CredentialWriter from "../credentials/writer";
 
 interface ProfileInfoDrawerProps {
   profile: ParticipantProfile;
@@ -12,8 +10,6 @@ interface ProfileInfoDrawerProps {
 }
 
 function ProfileInfoDrawer({ profile, onClose }: ProfileInfoDrawerProps) {
-  const tuyau = useTuyau();
-
   return (
     <Drawer
       defaultOpen={true}
@@ -26,36 +22,26 @@ function ProfileInfoDrawer({ profile, onClose }: ProfileInfoDrawerProps) {
       }}
     >
       <DrawerContent className="bg-enei-beige text-enei-blue absolute gap-4 p-4">
-        <div className="mx-auto flex w-1/4 flex-row items-center justify-between">
-          <User className="h-48 w-48"></User>
-          <div className="flex flex-col gap-1">
-            <Button asChild>
-              <Link
-                route="pages:profile.show"
-                params={{ slug: profile.slug ?? "" }}
-                target="_blank"
-              >
-                <p>Ir para o perfil</p>
-              </Link>
-            </Button>
-            <p className="">{`${profile.firstName} ${profile.lastName}`}</p>
-            <p className="">{`${profile.university} | ${profile.course}`}</p>
+        <div className="mx-auto flex w-full max-w-96 flex-col items-center gap-2">
+          <div className="relative flex flex-row items-center">
+            <User className="size-16" />
+            <div className="flex flex-col">
+              <p>
+                <Link
+                  className="after:absolute after:inset-0 hover:underline"
+                  route="pages:profile.show"
+                  params={{ slug: profile.slug ?? "" }}
+                  target="_blank"
+                >
+                  {profile.firstName} {profile.lastName}
+                </Link>
+              </p>
+              <p>{profile.slug}</p>
+            </div>
           </div>
-        </div>
 
-        <NFC
-          makeReadOnly
-          writeValue={{
-            records: [
-              {
-                recordType: "url",
-                data: tuyau.$url("pages:profile.show", {
-                  params: { slug: profile.slug ?? "" },
-                }),
-              },
-            ],
-          }}
-        />
+          <CredentialWriter slug={profile.slug ?? ""} />
+        </div>
       </DrawerContent>
     </Drawer>
   );
